requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.9,<2.10
PyJWT>=2.9.0,<2.10